% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pc.cov.structure.R
\name{pc.cov.structure}
\alias{pc.cov.structure}
\title{Estimate the periodically correlated covariance structure within a given window \eqn{k \in [-q,q]}}
\usage{
pc.cov.structure(X, Y, q, d, T)
}
\arguments{
\item{X}{first process}

\item{Y}{second process, if null then autocovariance of \code{X} is computed}

\item{q}{size of the window (covariances from \code{-q} to \code{q} will be computed)}

\item{d}{index in period between \eqn{0} and \eqn{T-1}}

\item{T}{period}
}
\value{
a time domain operator
}
\description{
Given two multivariate periodically correlated time series \eqn{X_t} and \eqn{Y_t} with period \eqn{T}
estimates the covariances of \eqn{Cov(X_{d+k} Y_d)} for \eqn{k \in [-q,q]} and some index \eqn{d}.
\code{\link{lagged.cov}} is used for the estimation at each lag.
}
\examples{
X = rar(100)
Y = rar(100)
pc.cov.structure(X,Y,3,0,3)
}

