% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pc.lagged.cov.R
\name{pc.lagged.cov}
\alias{pc.lagged.cov}
\title{Compute cross covariance with a given lag for periodically correlated time series}
\usage{
pc.lagged.cov(X, Y, lag, d, T)
}
\arguments{
\item{X}{first process}

\item{Y}{second process, if null then autocovariance of X is computed}

\item{lag}{the lag that we are interested in}

\item{d}{index in period}

\item{T}{period}
}
\value{
Covariance matrix
}
\description{
For a given multivariate stationary periodically correlated time series estimates a covarianve matrix
\eqn{C_{XY}^k_d = Cov(X_{d+k},Y_d)} using the formula
\deqn{\hat C_{XY}^k_d =  \frac{1}{[n/T]} \sum_{i=1}^{[n/T]} X_{k+i*T+d} Y_{k+d}'. }
}
\examples{
X = rar(100)
Y = rar(100)
pc.lagged.cov(X,Y)
}

