% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcdpca.inverse.R
\name{pcdpca.inverse}
\alias{pcdpca.inverse}
\title{Retrieve a process from given scores}
\usage{
pcdpca.inverse(Y, XI)
}
\arguments{
\item{Y}{scores process}

\item{XI}{principal components series}
}
\value{
Retrived process X
}
\description{
For given scores \code{Y} and dynamic principal components \code{XI}
retrive a series from which scores \code{Y} were calculated.
This procedure should be seen as the inverse of \code{\link{pcdpca.scores}}.
}
\references{
Kidzinski, Kokoszka, Jouzdani
Dynamic principal components of periodically correlated functional time series
Research report, 2016
}
\seealso{
\code{\link{pcdpca.scores}}, \code{\link{pcdpca}}
}
