% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcdpca.scores.R
\name{pcdpca.scores}
\alias{pcdpca.scores}
\title{Compute periodically correlated DPCA filters and scores}
\usage{
pcdpca.scores(X, PC)
}
\arguments{
\item{X}{multivariate time series}

\item{PC}{series of filters returned from pcdpca}
}
\description{
Compute periodically correlated DPCA filters and scores
}
\examples{
pcdpca.scores(X)
}
\keyword{pcdpca}

