% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcdpca.R
\name{pcdpca}
\alias{pcdpca}
\title{Compute periodically correlated DPCA filters and scores}
\usage{
pcdpca(X)
}
\arguments{
\item{X}{multivariate time series}
}
\description{
Compute periodically correlated DPCA filters and scores
}
\examples{
pcdpca(X)
}
\keyword{pcdpca}

