% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcdpca.scores.R
\name{pcdpca.scores}
\alias{pcdpca.scores}
\title{Compute periodically correlated DPCA scores, given the filters XI}
\usage{
pcdpca.scores(X, XI)
}
\arguments{
\item{X}{multivariate time series}

\item{XI}{series of filters returned from pcdpca}
}
\description{
Compute periodically correlated DPCA scores, given the filters XI
}
\keyword{pcdpca}
